laboratory work 1_3 update data type

+++ src/lab_1/lab_1_3.java
@@ -2,8 +2,8 @@ package lab_1;
 
 public class lab_1_3 {
     public static void printlab_1_3() {
-        int number = 10500;
-        int results;
+        double number = 10500;
+        double results;
         results = (number / 10) / 10;
         System.out.println("Лаб. 1-3: \n" +
                 "results = " + results + "\n");
